operator: move generic API functions from register.go to server.go
We are introducing new APIs: let's keep in the register.go package
only those functions specific to the register API only.

This commit just moves some functions from register.go to server.go.
No changes in the code.
